renamed 'src/renamte.ts' to 'src/workspace-symbol.ts'
