+++ src/utils/builtins.ts
@@ -625,4 +625,3 @@ const PossibleSubCommand = [
     "fab",
 ];
 const SubCommandSet = new Set(...PossibleSubCommand)
-
