+++ src/utils/completion-types.ts
@@ -6,91 +6,6 @@ import { FishCompletionItemKind } from './completion-strategy';
 //import {logger} from '../logger';
 import {execCommandDocs, execCommandType} from './exec';
 
-
-export const BuiltInList = [
-    "[",
-    "_",
-    "and",
-    "argparse",
-    "begin",
-    "bg",
-    "bind",
-    "block",
-    "break",
-    "breakpoint",
-    "builtin",
-    "case",
-    "cd",
-    "command",
-    "commandline",
-    "complete",
-    "contains",
-    "continue",
-    "count",
-    "disown",
-    "echo",
-    "else",
-    "emit",
-    "end",
-    "eval",
-    "exec",
-    "exit",
-    "false",
-    "fg",
-    "for",
-    "function",
-    "functions",
-    "history",
-    "if",
-    "jobs",
-    "math",
-    "not",
-    "or",
-    "path",
-    "printf",
-    "pwd",
-    "random",
-    "read",
-    "realpath",
-    "return",
-    "set",
-    "set_color",
-    "source",
-    "status",
-    "string",
-    "switch",
-    "test",
-    "time",
-    "true",
-    "type",
-    "ulimit",
-    "wait",
-    "while",
-];
-
-// You can generate this list by running `builtin --names` in a fish session
-// note that '.', and ':' are removed from the list because they do not contain
-// a man-page
-const BuiltInSET = new Set(BuiltInList);
-
-
-/**
- * line is an array of length 2 (Example's below). Builtins are retreived via:
- *          builtins --names 
- * builtins 
- *
- *     true	Do nothing, successfully
- *     while	Perform a command multiple times
- *
- * @param {string[]} line - a result from fish's builtin commandline completions
- * @return {boolean} - line is a completion for an builtin 
- */
-//export function isBuiltIn(line: string | [...string[]]): boolean {
-export function isBuiltIn(line: string): boolean {
-    return BuiltInSET.has(line)
-}
-
-
 export const escapeChars: {[char: string]: string} = {
     ['a']: 'escapes the alert character',
     ['b']: 'escapes the backspace character',
