+++ src/workspace-symbol.ts
